Extract Provider (and rename them) to their own file and move fixture(s) there.
