+++ features/messages/src/main/kotlin/io/element/android/features/messages/actionlist/ActionListStateProvider.kt
@@ -1,0 +1,47 @@
+/*
+ * Copyright (c) 2023 New Vector Ltd
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.element.android.features.messages.actionlist
+
+import androidx.compose.ui.tooling.preview.PreviewParameterProvider
+import io.element.android.features.messages.actionlist.model.TimelineItemAction
+import io.element.android.features.messages.timeline.aMessageEvent
+import kotlinx.collections.immutable.persistentListOf
+
+open class ActionListStateProvider : PreviewParameterProvider<ActionListState> {
+    override val values: Sequence<ActionListState>
+        get() = sequenceOf(
+            anActionListState(),
+            anActionListState().copy(target = ActionListState.Target.Loading(aMessageEvent())),
+            anActionListState().copy(
+                target = ActionListState.Target.Success(
+                    messageEvent = aMessageEvent(),
+                    actions = persistentListOf(
+                        TimelineItemAction.Reply,
+                        TimelineItemAction.Forward,
+                        TimelineItemAction.Copy,
+                        TimelineItemAction.Edit,
+                        TimelineItemAction.Redact,
+                    )
+                )
+            )
+        )
+}
+
+fun anActionListState() = ActionListState(
+    target = ActionListState.Target.None,
+    eventSink = {}
+)
