+++ features/messages/src/main/kotlin/io/element/android/features/messages/timeline/TimelineState.kt
@@ -20,7 +20,6 @@ import androidx.compose.runtime.Immutable
 import io.element.android.features.messages.timeline.model.TimelineItem
 import io.element.android.libraries.matrix.core.EventId
 import kotlinx.collections.immutable.ImmutableList
-import kotlinx.collections.immutable.persistentListOf
 
 @Immutable
 data class TimelineState(
@@ -28,11 +27,4 @@ data class TimelineState(
     val hasMoreToLoad: Boolean,
     val highlightedEventId: EventId?,
     val eventSink: (TimelineEvents) -> Unit
-)
-
-fun aTimelineState() = TimelineState(
-    timelineItems = persistentListOf(),
-    hasMoreToLoad = false,
-    highlightedEventId = null,
-    eventSink = {}
 )
