+++ features/messages/src/main/kotlin/io/element/android/features/messages/timeline/model/content/TimelineItemContentProvider.kt
@@ -1,0 +1,43 @@
+/*
+ * Copyright (c) 2023 New Vector Ltd
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.element.android.features.messages.timeline.model.content
+
+import androidx.compose.ui.tooling.preview.PreviewParameterProvider
+import org.matrix.rustcomponents.sdk.EncryptedMessage
+
+class TimelineItemContentProvider : PreviewParameterProvider<TimelineItemContent> {
+    override val values = sequenceOf(
+        TimelineItemEmoteContent(
+            body = "Emote",
+            htmlDocument = null
+        ),
+        TimelineItemEncryptedContent(
+            encryptedMessage = EncryptedMessage.Unknown
+        ),
+        // TODO MessagesTimelineItemImageContent(),
+        TimelineItemNoticeContent(
+            body = "Notice",
+            htmlDocument = null
+        ),
+        TimelineItemRedactedContent,
+        TimelineItemTextContent(
+            body = "Text",
+            htmlDocument = null
+        ),
+        TimelineItemUnknownContent,
+    )
+}
