+++ features/preferences/src/main/kotlin/io/element/android/features/preferences/root/PreferencesRootStateProvider.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022 New Vector Ltd
+ * Copyright (c) 2023 New Vector Ltd
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,6 +14,14 @@
  * limitations under the License.
  */
 
-package io.element.android.features.messages.timeline.model.content
+package io.element.android.features.preferences.root
 
-sealed interface TimelineItemContent
+import io.element.android.features.logout.aLogoutPreferenceState
+import io.element.android.features.rageshake.preferences.aRageshakePreferencesState
+import io.element.android.libraries.architecture.Async
+
+fun aPreferencesRootState() = PreferencesRootState(
+    logoutState = aLogoutPreferenceState(),
+    rageshakeState = aRageshakePreferencesState(),
+    myUser = Async.Uninitialized
+)
