+++ features/rageshake/src/main/kotlin/io/element/android/features/rageshake/crash/ui/CrashDetectionStateProvider.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 New Vector Ltd
+ * Copyright (c) 2022 New Vector Ltd
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,11 +14,9 @@
  * limitations under the License.
  */
 
-package io.element.android.libraries.designsystem.utils
+package io.element.android.features.rageshake.crash.ui
 
-import androidx.compose.ui.tooling.preview.PreviewParameterProvider
-
-open class BooleanPreviewParameterProvider : PreviewParameterProvider<Boolean> {
-    override val values: Sequence<Boolean>
-        get() = sequenceOf(false, true)
-}
+fun aCrashDetectionState() = CrashDetectionState(
+    crashDetected = false,
+    eventSink = {}
+)
