+++ features/rageshake/src/main/kotlin/io/element/android/features/rageshake/preferences/RageshakePreferencesStateProvider.kt
@@ -16,9 +16,19 @@
 
 package io.element.android.features.rageshake.preferences
 
-data class RageshakePreferencesState(
-    val isEnabled: Boolean,
-    val isSupported: Boolean,
-    val sensitivity: Float,
-    val eventSink: (RageshakePreferencesEvents) -> Unit,
+import androidx.compose.ui.tooling.preview.PreviewParameterProvider
+
+open class RageshakePreferencesStateProvider : PreviewParameterProvider<RageshakePreferencesState> {
+    override val values: Sequence<RageshakePreferencesState>
+        get() = sequenceOf(
+            aRageshakePreferencesState().copy(isEnabled = true, isSupported = true, sensitivity = 0.5f),
+            aRageshakePreferencesState().copy(isEnabled = true, isSupported = false, sensitivity = 0.5f),
+        )
+}
+
+fun aRageshakePreferencesState() = RageshakePreferencesState(
+    isEnabled = false,
+    isSupported = true,
+    sensitivity = 0.3f,
+    eventSink = {}
 )
